did all wishlist and cart functions with reducer

+++ src/context/cart-context.js
@@ -1,5 +1,7 @@
-import { createContext, useContext, useState } from "react";
+import { createContext, useContext, useState, useReducer, useEffect } from "react";
 import axios from "axios"
+import { cartReducer } from "../reducer/cart-reducer";
+import {token} from "../utils/token"
 
 const CartContext = createContext();
 
@@ -8,21 +10,24 @@ function CartProvider({children}) {
     const [myCart, setMyCart] = useState([])
 
     const addToCart = async (token, product) => {
-      try {
-        const res = await axios.post('/api/user/cart', {product}, {
-          headers: {
-            authorization: token
-          }
-        })
-        setMyCart(res.data.cart)
-      }
-  
-      catch(err) {
-        console.log(err)
+      if(product !== null) {
+        try {
+          const res = await axios.post('/api/user/cart', {product}, {
+            headers: {
+              authorization: token
+            }
+          })
+          setMyCart(res.data.cart)
+        }
+    
+        catch(err) {
+          console.log(err)
+        }
       }
     }
 
-    const deleteFromCart = async (token, productId) => {
+    const deleteFromCart = async (token, productId, deleteProduct) => {
+      if(productId !== null && deleteProduct===true) {
         try {
           const res = await axios.delete(`/api/user/cart/${productId}`, {
             headers: {
@@ -35,29 +40,56 @@ function CartProvider({children}) {
         catch(err) {
           console.log(err)
         }
+      }
     }
 
     const updateCartQuantity = async (token, productId, updateType) => {
-      try {
-        const res = await axios.post(`/api/user/cart/${productId}`, {
-          action: {
-            type: updateType,
-          }
-        }, {
-          headers: {
-            authorization: token
-          }
-        })
-        setMyCart(res.data.cart)
-      }
-    
-      catch(err) {
-        console.log(err)
+      if(productId !== null) {
+        try {
+          const res = await axios.post(`/api/user/cart/${productId}`, {
+            action: {
+              type: updateType,
+            }
+          }, {
+            headers: {
+              authorization: token
+            }
+          })
+          setMyCart(res.data.cart)
+        }
+      
+        catch(err) {
+          console.log(err)
+        }
       }
     }
 
+    const productInCart = (myCart, productId) => myCart.some((ele) => ele._id === productId)
+
+    const [state, cartDispatch] = useReducer(cartReducer, {
+      product: null,
+      productId: null,
+      deleteProduct: false,
+      updateType: null,
+      adder: false,
+      deleter: false,
+      updater: false
+    })
+
+    useEffect(() => {
+      addToCart(token, state.product)
+    }, [state.adder])
+
+    useEffect(() => {
+      deleteFromCart(token, state.productId, state.deleteProduct)
+    }, [state.deleter])
+
+    useEffect(() => {
+      updateCartQuantity(token, state.productId, state.updateType)
+    }, [state.updater])
+
     return (
-        <CartContext.Provider value={{myCart, deleteFromCart, addToCart, updateCartQuantity}}>
+        <CartContext.Provider value={{myCart, cartDispatch, state, productInCart}}>
             {children}
         </CartContext.Provider>
     )
