+++ src/context/wishlist-context.js
@@ -1,12 +1,21 @@
-import { createContext, useContext, useState } from "react";
+import { createContext, useContext, useState, useReducer, useEffect } from "react";
 import axios from "axios";
+import { token } from "../utils/token";
+import { wishlistReducer } from "../reducer/wishlist-reducer";
 
 const WishlistContext = createContext();
 
 function WishlistProvider({children}) {
 
+  const [state, dispatch] = useReducer(wishlistReducer, {
+    product: null,
+    productId: null
+  })
+
   const [myWishlist, setMyWishlist] = useState([])
+
   const addToWishlist = async (token, product) => {
+    if(product !== null) {
       try {
         const res = await axios.post('api/user/wishlist', {product}, {
           headers: {
@@ -19,8 +28,11 @@ function WishlistProvider({children}) {
       catch(err) {
         console.log(err)
       }
+    } 
   }
+
   const deleteFromWishlist = async (token, productId) => {
+    if(productId !== null) {
       try {
         const res = await axios.delete(`/api/user/wishlist/${productId}`, {
           headers: {
@@ -33,9 +45,21 @@ function WishlistProvider({children}) {
       catch(err) {
         console.log(err)
       }
+    }
   }
+
+  const productInWishlist = (myWishlist, productId) => myWishlist.some((ele) => ele._id === productId)
+
+  useEffect(() => {
+    addToWishlist(token, state.product)
+  }, [state.product])
+
+  useEffect(() => {
+    deleteFromWishlist(token, state.productId)
+  }, [state.productId])
+
   return (
-      <WishlistContext.Provider value={{addToWishlist, deleteFromWishlist, myWishlist}}>
+      <WishlistContext.Provider value={{ myWishlist, state, dispatch, productInWishlist}}>
           {children}
       </WishlistContext.Provider>
   )
